// src/features/items/components/showcase/components/inline/interfaces/template-props.interface.ts
import type { Props } from './props.interface';

export type TemplateProps = Props & {
  statusBadgeClassName: string;
  heroMetaClassName: string;
  starClassName: string;
  heroClassName: string;
  priceRowClassName: string;
  priceClassName: string;
  bodyClassName: string;
  sectionClassName: string;
  sectionLabelClassName: string;
  propertyGridClassName: string;
  propCardClassName: string;
  propLabelClassName: string;
  propValueClassName: string;
  detailTextClassName: string;
  emptyDescriptionClassName: string;
  priorityClassName: string;
};